adding

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,6 +1,5 @@
 #include "function_pointers.h"
 #include <stdlib.h>
-#include <stddef.h>
 /**
  * array_iterator - function executes other fumction of array
  * @size:size of the array
@@ -12,7 +11,7 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 	int i;
 	if (array != NULL && action != NULL)
 	{
-		for (i = 0; i < size; i++)
+		for (i = 0; i < (int)size; i++)
 		{
 			action(array[i]);
 		}
